+++ depi/CHANGELOG.md
@@ -1,3 +1,10 @@
+## 1.1.0
+
+- Add generator package `depi_generator`.
+- Add annotations for `depi_generator`.
+- Create builder class `ServiceContainer`.
+- Add `ModifiableServiceProvider` to modify services after the container is created.
+
 ## 1.0.0
 
 - Initial version.